The `APIClient` now operates on `URL` types instead of `Strings` where possible. Fixes #9. (#30)
The only place where `Strings` had to be used is the `url` property of a `PageItem`. Here it is possible that the given page item is related to weather (e.g. City Weather) where the `url` contains placeholders. Placeholders are not supported by the built in `URL` type.

Switched over to building `URLs` from `URLComponents` instead of `Strings` in order to avoid formatting errors.

Updated queries to force-unwrap `URLComponents`, making the code less error-tolerant.